+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [9.2.0] - Unreleased
 - Add Internal Note to Catalogue
+- Fix issue with using Internal Catalogues in Cohort Identification Configurations
 
 ## [9.1.2] - Unreleased
 - Automatically fetch user settings from previous versions of RDMP when installing the latest version